Refactor : 사용자에게 메세지를 보여주는 UI 클래스 분리
